refactor(dm): extract live-sub reconnect controller (fixes file-size gate)
nostrLiveDmSub.ts grew to 1016 lines (cap 1000) as the #1039 review added
the reconnect-refresh settle timer + arm-generation bookkeeping. Extract
the whole connection-lifecycle state machine — open/backoff-reconnect/
resume-rearm/settle-timer — into nostrLiveDmReconnect.ts
(createLiveDmReconnectController), a single-responsibility module the
1:1/kind-4/kind-16-17 decode+persist+surface logic now composes against
via a plain onEvent callback and a start/stopReconnecting/closeSubscription
API. nostrLiveDmSub.ts drops to 807 lines; no logic, ordering, or guard
changed — all 11 nostrLiveDmSubRearm.test.ts cases pass unmodified
(they exercise startLiveDmSubscription as a black box), and the full
suite (166/166, 1914 tests) is green.
